LPS-181478 Remove FF and unnecessary styles
