+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page_structure/components/StructureTreeNodeActions.js
@@ -45,7 +45,7 @@ import useHasRequiredChild from '../../../../../app/utils/useHasRequiredChild';
 
 export default function StructureTreeNodeActions({
 	item,
-	setEditingName,
+	setEditingNodeId,
 	visible,
 }) {
 	const [active, setActive] = useState(false);
@@ -75,13 +75,16 @@ export default function StructureTreeNodeActions({
 				aria-haspopup="true"
 				aria-label={Liferay.Language.get('options')}
 				className={classNames(
-					'page-editor__page-structure__tree-node__actions-button',
+					'ml-0 page-editor__page-structure__tree-node__actions-button',
 					{
 						'page-editor__page-structure__tree-node__actions-button--visible': visible,
 					}
 				)}
 				displayType="unstyled"
-				onClick={() => updateActive(!active)}
+				onClick={(event) => {
+					event.stopPropagation();
+					updateActive(!active);
+				}}
 				ref={alignElementRef}
 				size="sm"
 				title={Liferay.Language.get('options')}
@@ -102,7 +105,7 @@ export default function StructureTreeNodeActions({
 					<ActionList
 						item={item}
 						setActive={updateActive}
-						setEditingName={setEditingName}
+						setEditingNodeId={setEditingNodeId}
 						setOpenSaveModal={setOpenSaveModal}
 					/>
 				)}
@@ -110,7 +113,7 @@ export default function StructureTreeNodeActions({
 
 			{openSaveModal && (
 				<SaveFragmentCompositionModal
-					itemId={item.itemId}
+					itemId={item.id}
 					onCloseModal={() => setOpenSaveModal(false)}
 				/>
 			)}
@@ -118,9 +121,9 @@ export default function StructureTreeNodeActions({
 	);
 }
 
-const ActionList = ({item, setActive, setEditingName, setOpenSaveModal}) => {
+const ActionList = ({item, setActive, setEditingNodeId, setOpenSaveModal}) => {
 	const dispatch = useDispatch();
-	const hasRequiredChild = useHasRequiredChild(item.itemId);
+	const hasRequiredChild = useHasRequiredChild(item.id);
 	const selectItem = useSelectItem();
 	const widgets = useSelector((state) => state.widgets);
 
@@ -147,7 +150,7 @@ const ActionList = ({item, setActive, setEditingName, setOpenSaveModal}) => {
 				action: () => {
 					updateItemStyle({
 						dispatch,
-						itemId: item.itemId,
+						itemId: item.id,
 						selectedViewportSize,
 						styleName: 'display',
 						styleValue: isHidden ? 'block' : 'none',
@@ -190,7 +193,7 @@ const ActionList = ({item, setActive, setEditingName, setOpenSaveModal}) => {
 				action: () =>
 					dispatch(
 						duplicateItem({
-							itemId: item.itemId,
+							itemId: item.id,
 							selectItem,
 						})
 					),
@@ -202,7 +205,7 @@ const ActionList = ({item, setActive, setEditingName, setOpenSaveModal}) => {
 		if (canBeRenamed(item)) {
 			items.push({
 				action: () => {
-					setEditingName(true);
+					setEditingNodeId(item.id);
 				},
 				label: Liferay.Language.get('rename'),
 			});
@@ -217,7 +220,7 @@ const ActionList = ({item, setActive, setEditingName, setOpenSaveModal}) => {
 				action: () =>
 					dispatch(
 						deleteItem({
-							itemId: item.itemId,
+							itemId: item.id,
 							selectItem,
 						})
 					),
@@ -239,7 +242,7 @@ const ActionList = ({item, setActive, setEditingName, setOpenSaveModal}) => {
 		widgets,
 		setOpenSaveModal,
 		isHidden,
-		setEditingName,
+		setEditingNodeId,
 	]);
 
 	return (
