+++ modules/apps/layout/layout-content-page-editor-web/src/main/resources/META-INF/resources/page_editor/plugins/browser/components/page_structure/components/_StructureTreeNode.scss
@@ -21,19 +21,21 @@ html#{$cadmin-selector} {
 			}
 
 			&--active.treeview-link {
-				background-color: $cadmin-primary-l3;
-				border-radius: 4px;
 				box-shadow: inset 0 0 0 1px $cadmin-primary-l1;
 				font-weight: 600;
 			}
 
-			&--hovered.treeview-link {
+			&--hovered.treeview-link,
+			&--active.treeview-link {
 				background-color: $cadmin-primary-l3;
 				border-radius: 4px;
 			}
-			&--mapped#{&}--active.treeview-link,
-			&--mapped#{&}--hovered.treeview-link {
-				background-color: fade_out($cadmin-purple, 0.92);
+			&--mapped#{&}--active,
+			&--mapped#{&}--hovered {
+				&.treeview-link {
+					background-color: fade_out($cadmin-purple, 0.92);
+				}
+				
 			}
 
 			&--mapped#{&}--active.treeview-link {
@@ -102,33 +104,6 @@ html#{$cadmin-selector} {
 				&:focus {
 					box-shadow: none;
 				}
-			}
-
-			&--active {
-				font-weight: 600;
-			}
-
-			&--active,
-			&--hovered {
-				background-color: $cadmin-primary-l3;
-				border-radius: 4px;
-			}
-
-			&--active {
-				box-shadow: inset 0 0 0 1px $cadmin-primary-l1;
-			}
-
-			&--mapped#{&}--active,
-			&--mapped#{&}--hovered {
-				background-color: fade_out($cadmin-purple, 0.92);
-			}
-
-			&--mapped#{&}--active {
-				box-shadow: inset 0 0 0 1px $cadmin-purple;
-			}
-
-			&--master-item {
-				background-color: $cadmin-light-l1;
 			}
 
 			&__buttons--hidden > button {
