+++ portal-impl/src/portal.properties
@@ -6366,11 +6366,6 @@
     feature.flag.LPS-149256=false
 
     #
-    # Env: LIFERAY_FEATURE_PERIOD_FLAG_PERIOD__UPPERCASEL__UPPERCASEP__UPPERCASES__MINUS__NUMBER1__NUMBER5__NUMBER1__NUMBER6__NUMBER7__NUMBER8_
-    #
-    feature.flag.LPS-151678=false
-
-    #
     # Env: LIFERAY_FEATURE_PERIOD_FLAG_PERIOD__UPPERCASEL__UPPERCASEP__UPPERCASES__MINUS__NUMBER1__NUMBER5__NUMBER3__NUMBER1__NUMBER1__NUMBER7_
     #
     feature.flag.LPS-153117=false
